ZONE_DMA32: looks like working but has these flaws:
The zone hierarchy is HIGHMEM->NORMAL->DMA32->DMA.
As I test with 2G of memory the DMA is 2G-16M, DMA32 is 16M and
NORMAL and HIGHMEM is zero so I get DMA32 exhaust in first place
and all of my kmalloc(size,GFP_DMA32) gets allocated to DMA zone.

I want to stash this work for the record.